pyproject.toml
[tool.poetry]
name = "eho"
version = "0.1.0"
description = ""
authors = ["aziret"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.2"
django-jazzmin = "^3.0.1"
djangorestframework = "^3.16.0"
drf-spectacular = "^0.28.0"
drf-yasg = "^1.21.10"
pillow = "^11.2.1"
djoser = "^2.3.1"
async-timeout = "^5.0.1"
coreapi = "^2.3.3"
django-cors-headers = "^4.7.0"
django-debug-toolbar = "^5.1.0"
django-decouple = "^2.1"
markdown = "^3.8"
psycopg2-binary = "^2.9.10"
redis = "^5.2.1"
tzdata = "^2025.2"
whitenoise = "^6.9.0"
ruamel-yaml = "^0.18.10"
ruamel-yaml-clib = "^0.2.12"
gunicorn = "^23.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
